Set default for --max-hsps=1 Auto disable traceback

+++ src/basic/config.h
@@ -226,7 +226,7 @@ struct Config
 	size_t chaining_min_nodes;
 	bool fast_tsv;
 	unsigned target_parallel_verbosity;
-	size_t memory_limit;
+	double memory_limit;
 
 	Sensitivity sensitivity;
